Coil objects

CoilFields.Coils.Coil — Type

Coil{TT,GEOMETRY}

Geometry: Coil geometry, see subtypes of AbstractCoilGeometry

J: current

length: Coil resolution

If the Geometry is an AbstractArray, the coil it can be indexed with coil[i].

Also supports: eachindex and eltype.

CoilFields.Coils.CoilSet — Type

CoilSet{TT, GEOMETRY}

Store coils as a vector of object Coil

A CoilSet can be generated by using readcoilset or getcoilset from the QUASR database.

CoilFields.Coils.CompositeCoilSet — Type

CompositeCoilSet{TT,TGROUP<:Tuple}

Contains a Group of CoilSet objects.

Geometry